requirements.txt
# 🛡️ SECURE AI STUDIO - Python Dependencies
# Production-ready requirements with security focus

# Web Framework (for API/UI) - Essential
flask
flask-limiter
requests

# Security & Utilities - Essential
pydantic
tqdm
python-dotenv
boto3
cryptography
pycryptodome

# System Monitoring
psutil
GPUtil

# AI Model API Client - Essential for functionality
replicate

# Image Processing - Essential
Pillow
numpy

# Testing - Essential
pytest
pytest-cov
pytest-mock
coverage

# Development (optional)
black
flake8
mypy